reorganize routes to leverage templating feature
